+++ src/main/java/lotto/dto/LottoResponse.java
@@ -1,4 +1,4 @@
-package lotto.model.lotto;
+package lotto.dto;
 
 import java.util.Collections;
 import java.util.HashSet;
