+++ src/main/java/lotto/dto/WinningLottoResponse.java
@@ -1,4 +1,6 @@
-package lotto.model.lotto;
+package lotto.dto;
+
+import lotto.model.lotto.WinningLotto;
 
 import java.util.Collections;
 import java.util.HashSet;
