+++ src/main/java/lotto/model/lotto/Lotto.java
@@ -2,6 +2,7 @@ package lotto.model.lotto;
 
 import java.util.*;
 
+import lotto.dto.WinningLottoResponse;
 import lotto.model.message.LottoNumberExceptionMessage;
 import lotto.model.lotto.result.Rank;
 import lotto.model.lotto.result.WinningResult;
