+++ src/test/java/lotto/model/lotto/LottoTest.java
@@ -2,6 +2,7 @@ package lotto.model.lotto;
 
 import java.util.*;
 
+import lotto.dto.WinningLottoResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
